+++ applovin_max/android/src/main/java/com/applovin/applovin_max/AppLovinMAXNativeAdView.java
@@ -1,7 +1,12 @@
 package com.applovin.applovin_max;
 
 import android.content.Context;
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
 import android.graphics.Rect;
+import android.graphics.drawable.BitmapDrawable;
+import android.graphics.drawable.Drawable;
+import android.text.TextUtils;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.FrameLayout;
@@ -16,10 +21,14 @@ import com.applovin.mediation.nativeAds.MaxNativeAdLoader;
 import com.applovin.mediation.nativeAds.MaxNativeAdView;
 import com.applovin.sdk.AppLovinSdk;
 
+import java.io.InputStream;
+import java.net.URL;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.Executor;
+import java.util.concurrent.Executors;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 import androidx.annotation.NonNull;
@@ -28,6 +37,8 @@ import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.platform.PlatformView;
+
+import static com.applovin.sdk.AppLovinSdkUtils.runOnUiThread;
 
 public class AppLovinMAXNativeAdView
         implements PlatformView, MaxAdRevenueListener
@@ -38,6 +49,8 @@ public class AppLovinMAXNativeAdView
     private static final int BODY_VIEW_TAG            = 4;
     private static final int CALL_TO_ACTION_VIEW_TAG  = 5;
     private static final int ADVERTISER_VIEW_TAG      = 8;
+
+    private static final Executor cachingExecutor = Executors.newSingleThreadExecutor();
 
     private final Context       context;
     private final MethodChannel channel;
@@ -369,6 +382,7 @@ public class AppLovinMAXNativeAdView
         {
             // TODO: Update import when MAX SDK with API is released.
             // AppLovinSdkUtils.setImageUrl( icon.getUri(), iconView );
+            setImageUrl( icon.getUri().toString(), iconView );
         }
         else if ( icon.getDrawable() != null )
         {
@@ -524,5 +538,31 @@ public class AppLovinMAXNativeAdView
         }
 
         clickableViews.clear();
+    }
+
+    private void setImageUrl(final String imageUrl, final ImageView imageView)
+    {
+        if ( TextUtils.isEmpty( imageUrl ) ) return;
+        if ( imageView == null ) return;
+
+        cachingExecutor.execute( () -> {
+
+            try
+            {
+                InputStream inputStream = new URL( imageUrl ).openStream();
+                Bitmap imageData = BitmapFactory.decodeStream( inputStream );
+
+                runOnUiThread( () -> {
+                    Drawable imageDrawable = new BitmapDrawable( context.getResources(), imageData );
+                    imageView.setImageDrawable( imageDrawable );
+                } );
+
+                inputStream.close();
+            }
+            catch ( Throwable th )
+            {
+                AppLovinMAX.e( "Failed to fetch image: " + imageUrl + " because of: " + th );
+            }
+        } );
     }
 }